Move applyPendingMemberships from LocalGroup to LocalAddressBook
