+++ core/src/main/kotlin/at/bitfire/davdroid/resource/LocalGroup.kt
@@ -28,81 +28,8 @@ import at.bitfire.synctools.storage.contacts.ContactsBatchOperation
 import com.google.common.base.MoreObjects
 import java.util.LinkedList
 import java.util.Optional
-import java.util.logging.Logger
-import kotlin.jvm.optionals.getOrNull
 
 class LocalGroup : AndroidGroup, LocalAddress {
-
-    companion object {
-
-        private val logger: Logger
-            get() = Logger.getGlobal()
-
-        /**
-         * Processes all groups with non-null [GroupColumns.PENDING_MEMBERS]: the pending memberships
-         * are applied (if possible) to keep cached memberships in sync.
-         *
-         * @param addressBook    address book to take groups from
-         */
-        fun applyPendingMemberships(addressBook: LocalAddressBook) {
-            logger.info("Assigning memberships of contact groups")
-
-            addressBook.allGroups { group ->
-                val groupId = group.id!!
-                val pendingMemberUids = group.pendingMemberships.toMutableSet()
-                val batch = ContactsBatchOperation(addressBook.provider!!)
-
-                // required for workaround for Android 7 which sets DIRTY flag when only meta-data is changed
-                val changeContactIDs = HashSet<Long>()
-
-                // process members which are currently in this group, but shouldn't be
-                for (currentMemberId in addressBook.getContactIdsByGroupMembership(groupId)) {
-                    val uid = addressBook.getContactUidFromId(currentMemberId) ?: continue
-
-                    if (!pendingMemberUids.contains(uid)) {
-                        logger.fine("$currentMemberId removed from group $groupId; removing group membership")
-                        val currentMember = addressBook.findContactById(currentMemberId)
-                        currentMember.removeGroupMemberships(batch)
-
-                        // Android 7 hack
-                        changeContactIDs += currentMemberId
-                    }
-
-                    // UID is processed, remove from pendingMembers
-                    pendingMemberUids -= uid
-                }
-                // now pendingMemberUids contains all UIDs which are not assigned yet
-
-                // process members which should be in this group, but aren't
-                for (missingMemberUid in pendingMemberUids) {
-                    val missingMember = addressBook.findContactByUid(missingMemberUid)
-                    if (missingMember == null) {
-                        logger.warning("Group $groupId has member $missingMemberUid which is not found in the address book; ignoring")
-                        continue
-                    }
-
-                    logger.fine("Assigning member $missingMember to group $groupId")
-                    missingMember.addToGroup(batch, groupId)
-
-                    // Android 7 hack
-                    changeContactIDs += missingMember.id!!
-                }
-
-                addressBook.dirtyVerifier.getOrNull()?.let { verifier ->
-                    // workaround for Android 7 which sets DIRTY flag when only meta-data is changed
-                    changeContactIDs
-                        .map { id -> addressBook.findContactById(id) }
-                        .forEach { contact ->
-                            verifier.updateHashCode(contact, batch)
-                        }
-                }
-
-                batch.commit()
-            }
-        }
-
-    }
-
 
     override var scheduleTag: String?
         get() = null
